requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
chardet==4.0.0
deep-daze==0.8.1
discord==1.0.1
discord.py==1.6.0
einops==0.3.0
fire==0.4.0
ftfy==5.9
idna==3.1
imageio==2.9.0
multidict==5.1.0
numpy==1.20.2
Pillow==8.1.2
pycodestyle==2.7.0
python-dotenv==0.16.0
pytorch-ranger==0.1.1
regex==2021.3.17
rope==0.18.0
siren-pytorch==0.0.9
six==1.15.0
termcolor==1.1.0
toml==0.10.2
torch==1.8.1
torch-optimizer==0.1.0
torchvision==0.9.1
tqdm==4.59.0
typing-extensions==3.7.4.3
wcwidth==0.2.5
yarl==1.6.3
